requirements.txt
flask
flask-socketio
eventlet
playwright
google-generativeai
python-dotenv
customtkinter
Pillow